pyproject.toml
[project]
name = "meeting-transcriber"
version = "0.1.0"
description = "Automated audio transcription and analysis using Gemini API"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "PyYAML>=6.0",
    "anthropic>=0.40.0",
    "faster-whisper>=1.0.0",
    "ollama>=0.3.0",
]

[tool.setuptools]
py-modules = ["auto_transcribe", "reclassify_and_fix", "ondemand_transcribe", "pipeline", "config"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["utils/"]
